Created selectors for configFields
